Adding general Wigner d-function and allowing spin 3 kaon resonances. If we want spin 4, wigner function is ready, but we would need Blatt-Weisskopf barrier factor.

+++ pdfs/dalitz/include/DPWignerFunction.hh
@@ -9,7 +9,7 @@ class DPWignerFunction
     DPWignerFunction() {};
     ~DPWignerFunction() {};
 
-    virtual double function(double cosTheta, int m, int n) = 0;
+    virtual double function(double cosTheta, double m, double n) = 0;
 
   protected:
